refactor: [GNS-354] Extract StorageDeposit as common
