+++ src/repositories/chain-repository/types.ts
@@ -1,4 +1,5 @@
 import { Amount } from "@/types";
+import { StorageDeposit } from "@/models/storage-deposit-model";
 
 export interface IChainRepository {
   getTokenSupply(): Promise<TokenSupplyInfo>;
@@ -8,6 +9,8 @@ export interface IChainRepository {
   getValidatorInfos(chainId: string): Promise<ValidatorInfo[]>;
 
   getStoragePrice(): Promise<Amount | null>;
+
+  getTotalStorageDeposit(): Promise<StorageDeposit | null>;
 }
 
 export interface TokenSupplyInfo {
